api/requirements.txt
fastapi[standard]>=0.110.0
langchain>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.26
pydantic>=2.6.0
python-dotenv>=1.0.1
uvicorn>=0.28.0
